requirements.txt
versiontools

flask

flask-migrate

flask-login

flask-cors

Flask-SQLAlchemy

Flask-SocketIO

Jinja2

MarkupSafe

Werkzeug

gevent

gevent-socketio

gevent-websocket

greenlet

itsdangerous

ujson

python-dotenv